requirements.txt
streamlit>=1.28
google-genai>=0.1.0
pandas>=2.0
python-dotenv>=1.0
